The goal of this project is to design an SDRAM controller that allows SDRAM memory to be interfaced  with a microprocessor having only asynchronous memory support.
The goal of this project is to design an SDRAM controller that allows SDRAM memory to be interfaced with a
microprocessor having only asynchronous memory support. The requirement is that support for a single
MT48LC8M16A2 be provided. Interfacing this SDRAM with any 32-bit data processor provided SDRAM
memory support is not already provided for that processor. In this project we have taken 80386DX to interface
with.